+++ pacman-server/src/main.ts
@@ -9,7 +9,7 @@ import store from './store'
 import MainBoard from '@/components/MainBoard.vue'
 import TestBoard from '@/components/TestingBoard.vue'
 import ScoreBoard from './components/ScoreBoard.vue'
-import Admin from './views/Admin.vue'
+import Admin from './components/Admin.vue'
 
 Vue.config.productionTip = false;
